% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargraph_plot.R
\name{bargraph_plot}
\alias{bargraph_plot}
\title{Creating a bar graph}
\usage{
bargraph_plot(data, x_data, y_data, color, title, x_title, y_title)
}
\arguments{
\item{data}{Where the data is selected from (data frame)}

\item{x_data}{The column of the data that serves as the x-axis}

\item{y_data}{The column of the data that serves as the y-axis}

\item{color}{The column of data that will determine bar colors}

\item{title}{The title of the bar graph that will be above the data (string)}

\item{x_title}{The title of the x-axis (string)}

\item{y_title}{The title of the y-axis (string)}
}
\value{
The visual bar graph
}
\description{
This function allows one to create a bar graph with colored bars along with editable title and axes titles.
Returns the visual graph
}
