% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Creating a linear model of data}
\usage{
linear_model(data, x_var, y_var)
}
\arguments{
\item{data}{Where the data is selected from (data frame)}

\item{x_var}{The name of the column in quotes of numeric data that serves as the x-variable (string)}

\item{y_var}{The name of the column in quotes of numeric data that serves as the y-variable (string)}
}
\value{
The summary of the linear model information
}
\description{
This function allows one to create a linear model of a data frame.
}
