% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_calc.R
\name{median_calc}
\alias{median_calc}
\title{Calculating the median}
\usage{
median_calc(data, column)
}
\arguments{
\item{data}{Where the data is selected from (data frame)}

\item{column}{The name of the column containing the numerical data (string)}
}
\value{
The median of the data from the column selected
}
\description{
This function allows data to be described in terms of median for a data set that is based on averages taken from individuals.
Returns median value
}
