% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_data.R
\name{selecting_data}
\alias{selecting_data}
\title{Selecting specific data columns for use}
\usage{
selecting_data(data, columns)
}
\arguments{
\item{data}{Where data is selected from (data frame)}

\item{columns}{Name(s) of desired column(s) from initial data (string)}
}
\value{
The selected columns (data frame)
}
\description{
This function allows one to clean their data by selecting specific columns out of their data frame to work with and visualize.
Returns the selected columns as a new data frame
}
