% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_sodium.R
\name{splot_sodium}
\alias{splot_sodium}
\title{Creates a scatterplot}
\usage{
splot_sodium(data, x_data, y_data, color, title, x_title, y_title)
}
\arguments{
\item{data}{Where the data is selected from (data frame)}

\item{x_data}{The column of the data that serves as the x-axis}

\item{y_data}{The column of the data that serves as the y-axis}

\item{color}{The column of data that will determine point colors}

\item{title}{The title of the scatterplot that will be above the data (string)}

\item{x_title}{The title of the x-axis (string)}

\item{y_title}{The title of the y-axis (string)}
}
\value{
The visual scatterplot from input data
}
\description{
This function allows one to create a scatterplot of a data frame with the color of the points being determined based off of a categorical variable in the data frame.
Returns the visual scatterplot
}
