% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarizes data}
\usage{
summarize_data(data, continuous_var)
}
\arguments{
\item{data}{Where the data is selected from (data frame)}

\item{continuous_var}{The name of the numeric column being summarized (string)}
}
\value{
A data frame with the summary of the data based on the grouping variable
}
\description{
This function allows one to summarize a continuous variable of a data frame. By doing this, one can compare the continuous data trends of the data frame.
Returns a data frame of the summary, including a mean, minimum, and maximum values.
}
